+++ src/test/03_method/main.cpp
@@ -31,7 +31,7 @@ struct Vec {
 
 int main() {
 	{ // register Vec
-		ReflMngr::Instance().RegisterTypeAuto<Vec>();
+		ReflMngr::Instance().RegisterType<Vec>();
 		ReflMngr::Instance().AddConstructor<Vec, float, float>();
 		ReflMngr::Instance().AddField<&Vec::x>("x");
 		ReflMngr::Instance().AddField<&Vec::y>("y");