+++ src/test/04_enum/main.cpp
@@ -12,7 +12,7 @@ enum class Color {
 };
 
 int main() {
-	ReflMngr::Instance().RegisterTypeAuto<Color>();
+	ReflMngr::Instance().RegisterType<Color>();
 	ReflMngr::Instance().AddField<Color::Red>("Red");
 	ReflMngr::Instance().AddField<Color::Green>("Green");
 	ReflMngr::Instance().AddField<Color::Blue>("Blue");