+++ src/test/05_overload/main.cpp
@@ -26,7 +26,7 @@ struct Vec {
 
 int main() {
 	{ // register Vec
-		ReflMngr::Instance().RegisterTypeAuto<Vec>();
+		ReflMngr::Instance().RegisterType<Vec>();
 		ReflMngr::Instance().AddField<&Vec::x>("x");
 		ReflMngr::Instance().AddField<&Vec::y>("y");
 		ReflMngr::Instance().AddMethod<MemFuncOf<Vec& (const Vec&)>::get(&Vec::operator+=)>(StrIDRegistry::Meta::operator_assign_add);