+++ src/test/10_dynamic/main.cpp
@@ -14,7 +14,7 @@ enum class Color {
 int main() {
 	ReflMngr::Instance().AddDynamicField<const std::string>(GlobalID, "author", "Ubpa");
 	ReflMngr::Instance().AddDynamicField<const Color>(GlobalID, "theme", Color::Red);
-	ReflMngr::Instance().RegisterTypeAuto<int>();
+	ReflMngr::Instance().RegisterType<int>();
 	ReflMngr::Instance().AddDynamicField<const size_t>(TypeID_of<int>, "bits", sizeof(int) * 8);
 
 	std::cout << ReflMngr::Instance().RVar(GlobalID, "author") << std::endl;