+++ src/test/12_Meta/main.cpp
@@ -54,7 +54,7 @@ struct Vec {
 };
 
 int main() {
-	ReflMngr::Instance().RegisterTypeAuto<Vec>();
+	ReflMngr::Instance().RegisterType<Vec>();
 	ReflMngr::Instance().AddConstructor<Vec, float, float>();
 	ReflMngr::Instance().AddField<&Vec::x>("x");
 	ReflMngr::Instance().AddField<&Vec::y>("y");