+++ src/test/14_pch/B.cpp
@@ -3,7 +3,7 @@
 #include <UDRefl/UDRefl.h>
 
 void RegisterB() {
-	Ubpa::UDRefl::Mngr->RegisterTypeAuto<B>();
+	Ubpa::UDRefl::Mngr->RegisterType<B>();
 	Ubpa::UDRefl::Mngr->AddField<&B::v0>("v0");
 	Ubpa::UDRefl::Mngr->AddField<&B::v1>("v1");
 	Ubpa::UDRefl::Mngr->AddField<&B::v2>("v2");