+++ src/test/15_serializer/Vector.cpp
@@ -3,6 +3,6 @@
 #include <UDRefl/UDRefl.h>
 
 void RegisterVector() {
-	Ubpa::UDRefl::Mngr->RegisterTypeAuto<Vector>();
+	Ubpa::UDRefl::Mngr->RegisterType<Vector>();
 	Ubpa::UDRefl::Mngr->AddField<&Vector::data>("data");
 }