+++ src/test/16_container/main.cpp
@@ -5,7 +5,7 @@ using namespace Ubpa;
 using namespace Ubpa::UDRefl;
 
 int main() {
-	Mngr->RegisterTypeAuto<std::vector<size_t>>();
+	Mngr->RegisterType<std::vector<size_t>>();
 	SharedObject vec = Mngr->MakeShared(TypeID_of<std::vector<size_t>>);
 	for (size_t i = 0; i < 5; i++)
 		vec->push_back(i);